+++ source/pml/mllnnmts.c
@@ -16,16 +16,16 @@
 /* MAIN - a sample program */
 
 int main(int c, char **v)
-   {REAL *x, *y, result, sign;
+   {double *x, *y, result, sign;
     int i;
     int npoints = 10;
 
-    x = FMAKE_N(REAL, npoints, "MLLNNMTS.C:x");
-    y = FMAKE_N(REAL, npoints, "MLLNNMTS.C:y");
+    x = FMAKE_N(double, npoints, "MLLNNMTS.C:x");
+    y = FMAKE_N(double, npoints, "MLLNNMTS.C:y");
 
     for (i = 0; i < npoints; i++)
-        {x[i] = (REAL) i;
-         y[i] = (REAL) i;};
+        {x[i] = (double) i;
+         y[i] = (double) i;};
 
 /* L1 norm */
     result = PM_lnnorm(y, x, npoints, 1);
@@ -45,8 +45,8 @@ int main(int c, char **v)
 /* alternating signs */
     sign = 1.0;
     for (i = 0; i < npoints; i++, sign *= -1.0)
-        {y[i] = (REAL)i * sign;
-         x[i] = (REAL)i;}
+        {y[i] = (double) i * sign;
+         x[i] = (double) i;}
 
 /* L1 norm */
     result = PM_lnnorm(y, x, npoints, 1);
